refactor: WALLET-1353 — extract shared cancel-requests core
